Organise packaging a bit better
