+++ game/plugins/src/main/kotlin/gg/rsmod/plugins/content/skills/smithing/action/SmithingAction.kt
@@ -1,4 +1,4 @@
-package gg.rsmod.plugins.content.skills.smithing
+package gg.rsmod.plugins.content.skills.smithing.action
 
 import gg.rsmod.game.fs.def.ItemDef
 import gg.rsmod.game.model.queue.QueueTask
